pyproject.toml
[project]
name = "nonebot-plugin-qqtophone"
version = "0.2.1"
authors = [{ name = "StillMisty", email = "1772797842@qq.com" }]
description = "基于 Nonebot2，调用 API 获取对应 QQ 号码的信息。"
readme = "README.md"
license.file = "LICENSE"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Topic :: Communications :: Chat",
]
keywords = ["nonebot2", "nonebot", "plugin"]
dependencies = ["httpx>=0.27", "nonebot-adapter-onebot>=2.3", "nonebot2>=2.3"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project.urls]
Homepage = "https://github.com/StillMisty/nonebot_plugin_qqtophone"
